Inhalers in Pharmacy Shops (#456)
## About The Pull Request
Adds inhalers and refill containers to pharmacy shops, allowing anyone
with the asthma quirk to not constantly end up dead.
## Why It's Good For The Game
QOL if we decide to keep the Asthma negative Quirk. Offers a way for
asthmatic characters to not just die.
## Changelog
:cl:
add: Added Asthma inhalers and refill canisters for them to pharmacy
shops.
/:cl: